Move codegen test to an integration test
