+++ components/lark-codegen/Cargo.toml
@@ -6,3 +6,4 @@ edition = "2018"
 
 [dependencies]
 lark-mir = { path = "../lark-mir" }
+tempfile = "2.0"
